requirements.txt
Flask==3.1.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
marshmallow==3.21.1
python-dotenv==1.0.1
PyMySQL==1.1.1
ultralytics==8.4.23
opencv-python==4.13.0.92
alembic==1.18.4
bidict==0.23.1
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
filelock==3.25.2
Flask==3.1.3
Flask-Migrate==4.0.7
Flask-SocketIO==5.6.1
Flask-SQLAlchemy==3.1.1
fonttools==4.62.1
fsspec==2026.2.0
greenlet==3.3.2
h11==0.16.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.5.0
Mako==1.3.10
MarkupSafe==3.0.3
marshmallow==3.21.1
matplotlib==3.10.8
mpmath==1.3.0
MyApplication==0.1.0
networkx==3.6.1
numpy==2.4.3
opencv-python==4.13.0.92
packaging==26.0
pillow==12.1.1
polars==1.39.2
polars-runtime-32==1.39.2
psutil==7.2.2
PyMySQL==1.1.1
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.13.1
python-socketio==5.16.1
PyYAML==6.0.3
requests==2.32.5
scipy==1.17.1
simple-websocket==1.1.0
six==1.17.0
SQLAlchemy==2.0.48
sympy==1.14.0
torch==2.10.0
torchvision==0.25.0
typing_extensions==4.15.0
ultralytics==8.4.23
ultralytics-thop==2.0.18
urllib3==2.6.3
Werkzeug==3.1.6
wsproto==1.3.2
Flask-Login
google-genai==1.69.0
google-api-core==2.30.1
fastapi==0.114.0
uvicorn==0.24.0
python-multipart==0.0.10